refactor: 테마 자원 삭제 및 추가가 관리자만 가능하도록 인가조건 구현
